refactor(*): move code to subfolders
