+++ myapp.vcxproj
@@ -130,27 +130,27 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="ambient_light.h" />
-    <ClInclude Include="camera.h" />
-    <ClInclude Include="checkerboard_material.h" />
+    <ClInclude Include="src\lights\ambient_light.h" />
+    <ClInclude Include="src\camera.h" />
+    <ClInclude Include="src\materials\checkerboard_material.h" />
     <ClInclude Include="cl\tools.cl" />
-    <ClInclude Include="color.h" />
-    <ClInclude Include="directional_light.h" />
-    <ClInclude Include="glass_material.h" />
-    <ClInclude Include="scene_manager.h" />
-    <ClInclude Include="specular_material.h" />
-    <ClInclude Include="normal_material.h" />
-    <ClInclude Include="point_light.h" />
-    <ClInclude Include="diffuse_material.h" />
-    <ClInclude Include="hittable.h" />
-    <ClInclude Include="light.h" />
-    <ClInclude Include="material.h" />
+    <ClInclude Include="src\color.h" />
+    <ClInclude Include="src\lights\directional_light.h" />
+    <ClInclude Include="src\materials\glass_material.h" />
+    <ClInclude Include="src\scene_manager.h" />
+    <ClInclude Include="src\materials\specular_material.h" />
+    <ClInclude Include="src\materials\normal_material.h" />
+    <ClInclude Include="src\lights\point_light.h" />
+    <ClInclude Include="src\materials\diffuse_material.h" />
+    <ClInclude Include="src\hittable.h" />
+    <ClInclude Include="src\lights\light.h" />
+    <ClInclude Include="src\materials\material.h" />
     <ClInclude Include="myapp.h" />
-    <ClInclude Include="object3d.h" />
-    <ClInclude Include="plane.h" />
-    <ClInclude Include="ray.h" />
-    <ClInclude Include="scene.h" />
-    <ClInclude Include="sphere.h" />
+    <ClInclude Include="src\3d-objects\object3d.h" />
+    <ClInclude Include="src\3d-objects\primitives\plane.h" />
+    <ClInclude Include="src\ray.h" />
+    <ClInclude Include="src\scene.h" />
+    <ClInclude Include="src\3d-objects\primitives\sphere.h" />
     <ClInclude Include="template\common.h" />
     <ClInclude Include="template\precomp.h" />
   </ItemGroup>
