+++ myapp.vcxproj.filters
@@ -17,61 +17,61 @@
     <ClInclude Include="cl\tools.cl">
       <Filter>template\cl</Filter>
     </ClInclude>
-    <ClInclude Include="camera.h">
+    <ClInclude Include="src\camera.h">
       <Filter>src</Filter>
     </ClInclude>
-    <ClInclude Include="scene.h">
+    <ClInclude Include="src\scene.h">
       <Filter>src</Filter>
     </ClInclude>
-    <ClInclude Include="sphere.h">
+    <ClInclude Include="src\3d-objects\primitives\sphere.h">
       <Filter>src\3d-objects\primitives</Filter>
     </ClInclude>
-    <ClInclude Include="object3d.h">
+    <ClInclude Include="src\3d-objects\object3d.h">
       <Filter>src\3d-objects</Filter>
     </ClInclude>
-    <ClInclude Include="ray.h">
+    <ClInclude Include="src\ray.h">
       <Filter>src</Filter>
     </ClInclude>
-    <ClInclude Include="color.h">
+    <ClInclude Include="src\color.h">
       <Filter>src</Filter>
     </ClInclude>
-    <ClInclude Include="plane.h">
+    <ClInclude Include="src\3d-objects\primitives\plane.h">
       <Filter>src\3d-objects\primitives</Filter>
     </ClInclude>
-    <ClInclude Include="hittable.h">
+    <ClInclude Include="src\hittable.h">
       <Filter>src</Filter>
     </ClInclude>
-    <ClInclude Include="light.h">
-      <Filter>src\light</Filter>
+    <ClInclude Include="src\lights\light.h">
+      <Filter>src\lights</Filter>
     </ClInclude>
-    <ClInclude Include="material.h">
+    <ClInclude Include="src\materials\material.h">
       <Filter>src\materials</Filter>
     </ClInclude>
-    <ClInclude Include="diffuse_material.h">
+    <ClInclude Include="src\materials\diffuse_material.h">
       <Filter>src\materials</Filter>
     </ClInclude>
-    <ClInclude Include="ambient_light.h">
-      <Filter>src\light</Filter>
+    <ClInclude Include="src\lights\ambient_light.h">
+      <Filter>src\lights</Filter>
     </ClInclude>
-    <ClInclude Include="normal_material.h">
+    <ClInclude Include="src\materials\normal_material.h">
       <Filter>src\materials</Filter>
     </ClInclude>
-    <ClInclude Include="checkerboard_material.h">
+    <ClInclude Include="src\materials\checkerboard_material.h">
       <Filter>src\materials</Filter>
     </ClInclude>
-    <ClInclude Include="specular_material.h">
+    <ClInclude Include="src\materials\specular_material.h">
       <Filter>src\materials</Filter>
     </ClInclude>
-    <ClInclude Include="point_light.h">
-      <Filter>src\light</Filter>
+    <ClInclude Include="src\lights\point_light.h">
+      <Filter>src\lights</Filter>
     </ClInclude>
-    <ClInclude Include="glass_material.h">
+    <ClInclude Include="src\materials\glass_material.h">
       <Filter>src\materials</Filter>
     </ClInclude>
-    <ClInclude Include="directional_light.h">
-      <Filter>src\light</Filter>
+    <ClInclude Include="src\lights\directional_light.h">
+      <Filter>src\lights</Filter>
     </ClInclude>
-    <ClInclude Include="scene_manager.h">
+    <ClInclude Include="src\scene_manager.h">
       <Filter>src</Filter>
     </ClInclude>
   </ItemGroup>
@@ -102,11 +102,11 @@
     <Filter Include="src\3d-objects\primitives">
       <UniqueIdentifier>{95d0b961-1267-45c7-b176-2a21c90a9731}</UniqueIdentifier>
     </Filter>
-    <Filter Include="src\light">
-      <UniqueIdentifier>{e67bc880-1157-42aa-82ea-2bacd299d0bb}</UniqueIdentifier>
-    </Filter>
     <Filter Include="src\materials">
       <UniqueIdentifier>{4f15d112-c6fb-463f-b21e-4167d2a70cbc}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="src\lights">
+      <UniqueIdentifier>{e67bc880-1157-42aa-82ea-2bacd299d0bb}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
 </Project>
