+++ src/3d-objects/object3d.h
