+++ src/camera.h
