+++ src/lights/ambient_light.h
