+++ src/lights/point_light.h
