+++ src/materials/checkerboard_material.h
