+++ src/materials/diffuse_material.h
