+++ src/ray.h
@@ -7,8 +7,7 @@ namespace RaytracingRenderer {
 	{
     public:
         Ray(): orig(float3(0,0,0)), dir(float3(0,0,1)) {}
-        Ray(const float3& origin, const float3& direction)
-            : orig(origin), dir(direction) {}
+        Ray(const float3& origin, const float3& direction): orig(origin), dir(direction) {}
 
         float3 at(float t) const {
             return orig + t * dir;
