+++ src/scene_manager.h
