+++ template/precomp.h
@@ -1221,37 +1221,37 @@ public:
 };
 
 // General raytracing.
-#include "ray.h"
-#include "hittable.h"
+#include "src/ray.h"
+#include "src/hittable.h"
 
 // Materials
-#include "material.h"
-#include "diffuse_material.h"
-#include "normal_material.h"
-#include "checkerboard_material.h"
-#include "specular_material.h"
-#include "glass_material.h"
+#include "src/materials/material.h"
+#include "src/materials/diffuse_material.h"
+#include "src/materials/normal_material.h"
+#include "src/materials/checkerboard_material.h"
+#include "src/materials/specular_material.h"
+#include "src/materials/glass_material.h"
 
 // 3D Objects.
-#include "object3d.h"
-#include "sphere.h"
-#include "plane.h"
+#include "src/3d-objects/object3d.h"
+#include "src/3d-objects/primitives/sphere.h"
+#include "src/3d-objects/primitives/plane.h"
 
 // Camera.
-#include "camera.h"
+#include "src/camera.h"
 
 // Scene.
-#include "scene.h"
+#include "src/scene.h"
 
 // Lighting.
-#include "light.h"
-#include "ambient_light.h"
-#include "point_light.h"
-#include "directional_light.h"
+#include "src/lights/light.h"
+#include "src/lights/ambient_light.h"
+#include "src/lights/point_light.h"
+#include "src/lights/directional_light.h"
 
 // Color utility.
-#include "color.h"
+#include "src/color.h"
 
-#include "scene_manager.h"
+#include "src/scene_manager.h"
 
 // EOF
